+++ src/main/java/com/javaloping/homr/model/PropertyHistory.java
@@ -4,7 +4,7 @@ import javax.persistence.*;
 import java.util.Date;
 
 /**
- * Created by victor on 04/07/15.
+ * @author victormiranda@gmail.com
  */
 @Entity
 @Table(name = "property_histories")
@@ -12,7 +12,7 @@ public class PropertyHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "propertyId")
@@ -26,11 +26,11 @@ public class PropertyHistory {
     @Temporal(TemporalType.DATE)
     private Date modifyDate;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
