+++ src/main/java/com/javaloping/homr/model/User.java
@@ -5,14 +5,14 @@ import java.util.Date;
 import java.util.List;
 
 /**
- * Created by victor on 04/07/15.
+ * @author victormiranda@gmail.com
  */
 @Entity
 @Table(name = "users")
 public class User {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String username;
 
@@ -41,11 +41,11 @@ public class User {
 
     /** GETTERS AND SETTERS **/
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
